修改ad采样

+++ APP/app_temp.h
@@ -14,8 +14,8 @@ typedef struct _app_temp	{
 	OS_EVENT           *MSG_Q;
 	OS_EVENT           *lock;
 	TIM_HandleTypeDef *pTECPWM;
-	s16 target_t[PIDCTRL_NUM];//目标温度 0.1
-	s16 current_t[TEMP_ID_NUMS];//当前温度 0.1
+	s32 target_t[PIDCTRL_NUM];//目标温度 0.1
+	s32 current_t[TEMP_ID_NUMS];//当前温度 0.1
 	s16 duty[PIDCTRL_NUM];//tec pwm占空比
 }_app_temp_t;
